Make session.service(name) return a Future
In the C++ API, qi::Session::service(…) returns a
qi::FutureSync<qi::AnyObject>.

To mimic the synchronous usage in Java, Session.service(…) returned a
simple AnyObject (waiting for the future to complete). But that way,
it was not possible to use it asynchronously.

Instead, we should always return a Future, so that the API user can
either use it asynchronously or synchronously (by explicitly calling
get()).

Change-Id: I6183f94a8665ba1944f55d69666f5cb87331cb5c
Reviewed-on: http://gerrit.aldebaran.lan/69681
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>