Expose then()/andThen() to Java
Add then() and andThen() methods to the Future java class.

They accept a QiFunction as parameter, taking a Future as parameter (the
future on which then() or andThen() has been called) and returning a
Future.

Currently, contrary to the C++ version, the QiFunction *must* return a
Future. To overcome this limitation, we will add the possibility to
create a future from an immediate value (which is already available in
C++).

In the same way, the parameter differs from the C++ version of then()
which passes the future value as function parameter. In Java, it passes
the future itself.

Change-Id: I32e0dd5db1a0ecec9731bc17502a2271b4561fe9
Reviewed-on: http://gerrit.aldebaran.lan/69678
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: gerrit
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>