Expose a simplified version of Future.connect()
The C++ callback is just a function called once when the Future
finishes. Expose a similar interface in Java (with the callback being an
interface with exactly one method to be implemented by the client).

The previous Java callback implementation defined an interface with 3
methods (onSuccess/onFailure/onComplete). JNI is not the right place to
convert a call to get() into 3 different methods.

Change-Id: I72ff61b9eb7a6ac1ec74584b93963d90602b1082
Reviewed-on: http://gerrit.aldebaran.lan/69740
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Tested-by: epinault <epinault@aldebaran-robotics.com>