Make qi Future match java.util.concurrent.Future
Java users knows the standard Future API. The qi Future provides more
features, but their common part should match both the signature and the
semantics of the standard API.

To enforce this, make it implement the interface
java.util.concurrent.Future and make the implementation fit.

Consequently, we now declare and throw proper and standard exception
types:
 - java.util.concurrent.CancellationException on cancellation;
 - java.util.concurrent.ExecutionException on execution error;
 - java.util.concurrent.TimeoutException on timeout.

Note that there is a subtle semantic difference between what it means to
be "cancelled" for a java Future and for a qi Future:
 - a java Future is "cancelled" as soon as a call to cancel(…) returned
   true (i.e. cancel has been successfully requested);
 - a qi Future is "cancelled" when the cancel request has been handled.

Concretely, this means that a java Future expects that get() throws a
CancellationException in the following sample:

    if (future.cancel(…))
      future.get(); // must throw a CancellationException

This post-condition is not guaranteed by the native libqi, so we make
the implementation match the expected semantics.

Change-Id: Ia6e1e998eb0f8bb52f2f6587d4e7cd15646d7de6
Reviewed-on: http://gerrit.aldebaran.lan/69931
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Tested-by: epinault <epinault@aldebaran-robotics.com>