Remove unused mayInterrupt flag in native part
The inconsistency between the Java and native parts were fixed by
<http://gerrit.aldebaran.lan/#/c/70343/2>.

However, either we decide to expose the flag "mayInterrupted" to native
part just in case, so we initialize its value from the Java flag (we
don't hardcode "true"); either we remove the flag from the native part
because it is unused.

This commit implements the second solution.

Change-Id: Ia03268fe273b8348141582ae81a4477191e13509
Reviewed-on: http://gerrit.aldebaran.lan/70351
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>