Provide ExecutionException cause
The ExecutionException cause is not supposed to be null (it does not
even have a public constructor without a Throwable parameter).

Therefore, wrap the error message created by libqi in a QiException,
to be used as the ExecutionException cause.

It will also make possible to support strongly typed exceptions later
without impacting existing code (provided that they are subclasses of
QiException).

Change-Id: I6ab43a95f08d3a42d1d8697a6ae5afac0e342158
Reviewed-on: http://gerrit.aldebaran.lan/70366
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>