Add QiFunctionAdapter for then(…)/andThen(…)
Future then(…) and andThen(…) methods accept a QiFunction as parameter,
which maps the libqi C++ implementation.

It has a single method accepting a Future as parameter and returning a
Future. As a consequence, the user must use the future to retrieve
either the value, the error or the cancellation state.

For convenience, provide an adapter that dispatches these 3 cases into 3
different callbacks.

In particular, since a call to andThen(…) guarantees that the future has
a value, the future is not needed. Therefore:

    future.andThen(new QiFunction<Integer, AnyObject>() {
      @Override
      public Future<Integer> execute(Future<AnyObject> future) {
        AnyObject service = future.getValue();
        return service.call("addOne", 42);
      }
    });

can be replaced by:

    future.andThen(new QiFunctionAdapter<Integer, AnyObject>() {
      @Override
      public Future<Integer> handleResult(AnyObject service) {
        return service.call("addOne", 42);
      }
    });

Change-Id: I7019b1027afde319b56ccc94c7fb668d805c1ad4
Reviewed-on: http://gerrit.aldebaran.lan/70425
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: epinault <epinault@aldebaran-robotics.com>