Avoid data serialization in then(…)/andThen(…)
In order to use custom Java types as Promise/Future generic parameters
(povided that they are used only locally), we must avoid data
serialization. Otherwise, we may encounter this error:

> Cannot serialize return value: Unable to convert JObject to AnyValue

Calling native then(…) or andThen(…) eventually requires a call to
Future.unwrap(), which in turn leads get() to serialize data, that is
what we want to prevent.

Therefore, implement Java then(…) and andThen(…) in term of qi::Promise
and qi::Future, without calling native then(…) and andThen(…), so that
data serialization is not needed when returning an instance of a custom
Java class as the future value from a QiFunction.

Concretely, this makes this call work without errors, even if the Java
class X is not known by libqi:

  class X {}
  Future<Integer> future = Future.of(42);
  X x = future.then(new QiFunction<X, Integer>(Future<Integer> future) {
    @Override
    public Future<X> execute(Future<Integer> unused) {
      return Future.of(new X());
    }
  }).get();

Change-Id: Id7fdcb79890bdcc2776e194a021b1848b258a6ae
Reviewed-on: http://gerrit.aldebaran.lan/70504
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>