Expose a new way to connect to AnyObject signals
The existing signal connection mechanism had two problems:

 1. it required to pass the java callback signature as a String;
 2. since the connection involves remote calls, it blocked until the
    future associated were finished, while we want signal connection to
    return immediately.

To solve the first problem, pass the callback as a listener
(QiSignalListener) declaring exactly one method. There is a drawback
though: it cannot declare the parameter types, so its signature is:

    void onSignalReceived(Object... args);

But this implementation paves the way to implement in pure Java another
callback mechanism based on annotations, similar to the existing one but
without the need for explicit method signatures.

To solve the second problem, wrap the resulting future in a
QiSignalConnection. It is an opaque handle returned immediately, that
can be used to disconnect at any time.

Change-Id: I40a15654e2243fdc4deedc441c8863ff8e5fb558
Reviewed-on: http://gerrit.aldebaran.lan/70975
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: rvimont <rvimont@presta.aldebaran.com>