Convert tuples to custom structs
The @QiSlot mechanism introduced recently allowed to connect a signal to
a slot (a Java method) so that parameter types were explicit. However,
when one of the parameter were a Tuple, the client had to manage Tuple
manually.

This commit goes one step further, and provide automatic conversion
from Tuple to any object from a class annotated with @QiStruct.

Concretely, suppose that a signal is sent with a Tuple (struct) argument
having value { 42, "fourty-two" }:

    object.post("fire", Tuple.of(42, "fourty-two"));

We can connect a @QiSlot using strongly-typed struct as parameter:

    @QiStruct
    class Num {
      int numeric;
      String string;
    }

    @QiSlot
    void onNumReceivedRaw(Tuple tuple) {
      // raw tuple unchanged
      int n = tuple.get(0);    // 42
      String s = tuple.get(1); // "fourty-two"
    }

    @QiSlot
    void onNumReceived(Num num) {
      // raw tuple automatically mapped to a Num instance
      int n = num.numeric;     // 42
      String s = num.string;   // "fourty-two"
    }

    void connectSignals() {
      anyObject.connect("fire", this, "onNumReceivedRaw");
      anyObject.connect("fire", this, "onNumReceived");
    }

It also works recursively with lists and maps.

Change-Id: I1533a92c1198019e55c31e2f3f9135970e1f90f4
Reviewed-on: http://gerrit.aldebaran.lan/71095
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Tested-by: gerrit