Use a TypeToken to extract parameterized types
In order to convert tuples to structs (instances of classes annotated by
@QiStruct), we need to specify the target type. For example, if the
source is a list of tuples, we need a way to express that we want to
convert it to a list of Persons.

Since the generic parameter types are lost at runtime in Java, this is
not straightforward: we need a new type to encode the information for
each use.

Reuse the principle used by TypeToken in the GSON library:
<https://google.github.io/gson/apidocs/com/google/gson/reflect/TypeToken.html>

Change-Id: Ie5bdb8fd10945ebfe5222958f91c501947723e07
Reviewed-on: http://gerrit.aldebaran.lan/71284
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>