Hotfix undefined behavior on integer conversion
The variable 'value' is defined as qi::int64_t.

It is passed as a vararg argument to env->NewObject(…), so no implicit
conversion occurs. However, the implementation has no choice but
reinterpreting the argument as the expected type, leading to undefined
behavior, which has observable consequences on Android (arm).

To force the correct behavior, static_cast 'value' to the expected type.

Change-Id: Id9fc1e6be6ea69ad3ceb588f746bb7db1614de6b
Reviewed-on: http://gerrit.aldebaran.lan/71364
Reviewed-by: epinault <epinault@aldebaran-robotics.com>
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit