Improve JNI type system
Instances of supported types were used to lookup a jclass (by calling
GetObjectClass(…) on the instance) from a string, in order to avoid
calls to FindClass(…).

Instead, directly cache classes.

This greatly simplify usage: classes of supported types are directly
available from everywhere.

This is far more efficient, since it avoids lookup from a string and
retrieving the object class of the instance.

This allows to use interface types, like List and Map, so the user is
not limited to ArrayList and HashMap anymore.

Change-Id: I7740842fd651e2c65a938400212e130b82544810
Reviewed-on: http://gerrit.aldebaran.lan/71903
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>