Rename StructConverter to QiSerializer
The purpose of StructConverter os to convert structs (instances of
classes having @QiStruct annotation) from and to tuples.

But in fact, this conversion is only a particular case of general qi
serialization/deserialization, which convert custom objects from and to
objects of types supported by qi.

Therefore, rename it accordingly.

The new QiSerializer will then be refactored to support custom
conversions.

Change-Id: Ib703883318593fe9a3e8451177ddb447eb66ab57
Reviewed-on: http://gerrit.aldebaran.lan/72028
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>