Make QiSerializer more flexible
Instead of exposing a set of static methods, delegate the serialization
to an interface "Converter".

The serialization process iterates over the list of available
converters, and use the first that can handle the conversion.

Provide several default implementations, for lists, arrays, maps and
structs (instances of classes having the annotation @QiStruct).

The users can provide their own converters, that will be used if the
default ones don't match.

Change-Id: Ib836503163f3da535855543982316446954789e2
Reviewed-on: http://gerrit.aldebaran.lan/72029
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>