Introduce FutureFunction and its adapters
Future.connect() is a particular case of Future.then(), for which the
callback did not return any value.

Therefore, make it deprecated and provide tools to use then(…) instead.

The previous architecture was:

          +-----------------+        +------------+
          | Future.Callback |        | QiFunction |
          +-----------------+        +------------+
                   ^                        ^
                   |                        |
       +-----------------------+  +-------------------+
       | FutureCallbackAdapter |  | QiFunctionAdapter |
       +-----------------------+  +-------------------+

Future.Callback was used for connect(…), while QiFunction was used for
then(…)/andThen(…).

FutureCallbackAdapter and QiFunctionAdapter dispatched the execution to
3 different methods:
 - onResult(…)
 - onError(…)
 - onCancel(…)

The methods in the FutureCallbackAdapter version returned a Future,
while the one from the QiFunctionAdapter returned void.

Reorganize and simplify:

                    +----------------+
                    | FutureFunction | // replaces QiFunction
                    +----------------+
                           ^ ^
              .------------' '------------.
              |                           |
       +------------+               +------------+
       | QiFunction |               | QiCallback |
       +------------+               +------------+
       // replaces                  // replaces
       // QiFunctionAdapter         // FutureCallbackAdapter

In practice, FutureFunction is almost never used (albeit it is the more
general, visible in the method signatures). Therefore, reserve the
"short names" for the most used types.

It may seem surprising that the "void" version (QiCallback) implements
FutureFunction, but this is necessary to return a cancelled future in
case of cancellation. Moreover, it just provides an implementation
facility for something which is inherently a FutureFunction.

Change-Id: I19d151c3fd61f0d24f03f2fbb141982749a8eed7
Reviewed-on: http://gerrit.aldebaran.lan/72251
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>