Provide removeConnectionListener() on Session
Session only exposed addConnectionListener(), that directly connected
the listener to the native part.

Instead, keep the listeners registered by the user in a local list,
and connect one unique listener to the native part on the first call to
addConnectionListener(), that will delegate events to the local
listeners.

Thus, we can provide removeConnectionListener() easily.

Change-Id: I93f8b3e25380eb6a42f9c252d2414e9eb1b3b639
Reviewed-on: http://gerrit.aldebaran.lan/72512
Tested-by: gerrit
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>