Implement then(…)/andThen(…) in pure Java
In order to avoid data serialization in then(…)/andThen(…), a local
native version has been implemented¹ in term of qi::Promise and
qi::Future.

Since the objects it uses are exposed to Java, implement then(…) and
andThen(…) in pure Java.

This paves the way to provide the FutureCallbackType (Sync, Async, Auto)
as parameter in a consistant way.

¹ http://gerrit.aldebaran.lan/#/c/70504
  (commit having Change-Id Id7fdcb79890bdcc2776e194a021b1848b258a6ae)

Change-Id: Ia8e7c8289c54bb0119833924cc4d5f9bd232cf5e
Reviewed-on: http://gerrit.aldebaran.lan/72690
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>