Expose callback type to then/andThen
The callback type has been exposed to promises recently. Expose it also
to connect(…), then(…) and andThen(…).

Like in the native part, both the promise type and the type passed to
connect(…)/then(…)/andThen(…) influence the execution of callbacks.

The callback is executed either synchronously or asynchronously
(according to the thread that detected the finish of the future),
following these rules:

                            "then" type
                     +-----------------------+
                     | Sync  | Async | Auto  |
             +-------+-------+-------+-------+
             |  Sync | Sync    Async   Sync
     promise +-------+
        type | Async | Sync    ASync   Async
             +-------+
             |  Auto | Sync    Async     -
             +-------+

For example, in this code sample, the promise type is Auto (not
specified) while the "then" type is Sync.

    // T1 = Thread.currentThread.getId()
    QiCallback<Void> callback = …;
    final Promise<Void> promise = new Promise<>();
    promise.getFuture().then(callback, FutureCallbackType.Sync);
    new Thread(new Runnable() {
        @Override
        public void run() {
            // T2 = Thread.currentThread.getId()
            promise.setValue(null);
        }
    }).start();

As a result, the callback is executed *synchronously* according the
thread that detected the finish of the future, i.e. T2 (because it set
the promise value). In other words, the callback is indirectly called by
promise.setValue(…).

Change-Id: I38c8ac534d755659e184a1787b115c44a4262ae7
Reviewed-on: http://gerrit.aldebaran.lan/72691
Tested-by: gerrit
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>