Provide waitFor(futures)
Add a convenient method to make a future that waits for other futures to
finish. The resulting future returns the same value as the original, but
only when the other are finished.

Concretely:

    Future<T> future = …;
    Future<Void> conditionFulfilled = …;
    Future<T> result = myFuture.waitFor(conditionFulfilled);

Even if "future" may finish before "conditionFulfilled", "result" will
always finish once conditionFulfilled is finished successfully, with the
value of "future".

Change-Id: Ic52bf91e1b942e0674a940c0e0cb52ad2badf0b7
Reviewed-on: http://gerrit.aldebaran.lan/72826
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>