Read the future result atomically
A finished future may have 3 different states:
 - the future finished successfully: a value is available;
 - the future finished with an error: an error message is available;
 - the future has been cancelled.

However, once finished, its state may change to "cancelled" at any time
(after a call to future.cancel()).

Therefore, in general, this pattern, used internally, contains race
conditions:

    if (future.hasError())
      handleError(future.getErrorMessage()); // race condition
    else if (future.isCancelled())
      handleCancelled();
    else
      handleValue(future.getValue()); // race condition

To avoid the race conditions, the whole code were protected by the mutex
of the Future. However, this approach has a major drawback: the mutex is
kept locked while executing the if/else blocks.

In particular, some of these blocks set a promise, which may in turn
execute callbacks synchronously, with the mutex locked.

Instead, extract the result type and values to a FutureResult while the
mutex is locked, and use the FutureResult without mutex.

Change-Id: Ib214c1a21bf2bd917ff38d50986e67ac4c2bdf46
Reviewed-on: http://gerrit.aldebaran.lan/72853
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>