Declare throwing Throwable in callbacks
FutureFunction and its subclasses QiFunction/QiCallback methods declared
throwing Exception, but not Throwable.

This is annoying especially for onError(), because it prevents to
rethrow the received error (e.g. when we want to log and rethrow).

Therefore, allow the methods to throw any Throwable. This does not
impact existing code, because the new declaration is more general.

Change-Id: Ie22444736a12b2b04ddb403d6863552a7e2aace4
Reviewed-on: http://gerrit.aldebaran.lan/73215
Reviewed-by: rvimont <rvimont@presta.aldebaran.com>
Tested-by: gerrit
Reviewed-by: epinault <epinault@aldebaran-robotics.com>