Adding cancel propagation on Future.waitAll method #36641
The Future returned by this method was related to a promise that didn't have
a cancel callback. This callback has been added and when invoked, the cancel
propagation will be done to all the futures passed as argument to the waitAll
method. A test has been added to FutureTest as well.

Change-Id: Iee4936dbe8728c90dde7a363bb4d0b0ae9d039cf
Reviewed-on: http://gerrit.aldebaran.lan/76820
Reviewed-by: aroux <aroux@presta.aldebaran.com>
Reviewed-by: pduclos <pduclos@presta.aldebaran.com>
Reviewed-by: qjin <qjin@presta.aldebaran.com>
Tested-by: qjin <qjin@presta.aldebaran.com>
Reviewed-by: epinault <epinault@aldebaran-robotics.com>