Fix the forwarding of void futures
Futures were already properly forwarded, except the Void ones.
There was a specific case for Void and null return values, relying on the
advertised signature. This was wrong since through libqi you cannot tell
between a result or a future result.

Instead, the code now relies on the type of the actual value returned.
It involved the proper conversion of null return values, that are not
equivalent to qi::AnyReference(), but rather to
qi::AnyReference(qi::typeOf<void>()).

Relevant tests were added to demonstrate the fix.

Change-Id: Iacd53ac53e287193f5938fbddde450d423e81919