serialization: Adds an OptionalConverter.
Fixes #44342.

This patch adds the implementation of an `OptionalConverter` class and
adds it as one of the converters for the default `QiSerializer`.

This class allows the conversion between custom values contained in
`Optional` objects and values that the JNI layer of libqi-java handles
as part of the libqi type system.