java.Property: Inserts the setter in the Property constructor.
Fixes #45062
When we tried to construct a Property, it only did a memberwise setting.
Despites the value setter, it did not call the deserialization method needed by
some specific types, neither the testing part to avoid exceptions.

The solution was to call the value setter inside every constructor which take a
value.

Change-Id: I45b2aebddd330b5b8306174449ff235eaa895b1c
Reviewed-on: http://gerrit.aldebaran.lan/97034
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>