jni.future: Adds a jni method that converts a c++ future to a java future
Fixes #45711
There was a problem with functions that were creating a jlong pointer to a future,
which cast was incorrectly done by the jni part of the code

In this patch, jni function that was returning a jlong pointer now is returning a
jobject cast of a future, generated by a new converter function.

This new converter function, called toJavaFuture, accepts every kind of future,
excepts futures of a structure which signature is not known by libqi.

A memory leak has been corrected in java.Future's constructor.

Change-Id: I8b78f3875dcfab1ad45c92758bd9c43ac9cc3efb
Reviewed-on: http://gerrit2.aldebaran.lan/195
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>