java.EmbeddedTools: Synchronize threads on native libraries load. #45872
An UnsatisfiedLinkError occurred sometimes when more than one
thread tried to load and use native libraries. One thread did not
finish to load the native libraries when another concidered those
as loaded and try to use them.

The solution was to add a synchronization block to prevent multiple
threads to try loading the libraries at the same time, so they wait
so that it is guaranteed that the libraries are loaded when the
function returns, in any thread.

Change-Id: Ib67e4188d7a4e5bf3b407e7d3a851ad7eb483f2f
Reviewed-on: http://gerrit2.aldebaran.lan/278
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>