jni.jobjectconverter: Fixes memory leak by releasing local references. #45497
A crash occured when a function is called with, as argument, a list
containing a lot of elements, with the following error:
JNI ERROR (app bug): local reference table overflow (max=512)

Some of our conversion functions were creating new local references
in loops without releasing them inside the loops. This could lead
to a lot of references being held when converting for instance big
lists or maps of objects. The virtual machine has an arbitrary small
limit of local references being held at the same time, so this would
easily lead to crashes.

The solution was to modify them and make some functions returning new
local references that automatically releases the object at the end of
the scope in which it has been created.

Change-Id: I956c2f618f90088afbd47baef669f2df05f35a13
Reviewed-on: http://gerrit2.aldebaran.lan/330
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>