jni.test_common: Fixes incorrect access to EmbeddedTools member
There was a compatibility problem between two patches:
(1) jni.future: Adds a jni method that converts a c++ future to a java future
and
(2) java.EmbeddedTools: Synchronize threads on native libraries load. #45872

The (2) fix changed the type of a variable, embeddedLibrariesLoaded, a boolean
that prevents EmbeddedTools.java to invoke its libraries if it has already
been done.
This variable's type changed from AtomicBoolean to Boolean.

The (1) fix added a test that invokes those libraries first and then loads
the JVM. To prevent them being loading twice, it sets
embeddedLibrariesLoaded through jni.

Since (1) and (2) have been submitted in parallel, (1) cannot build anymore.
This fix changes the setting of embeddedLibrariesLoaded's definition in jni
to match with the type change included in (2).

Change-Id: Iff69f1ffb11583bd3c96f935c82c1bad0518c438
Reviewed-on: http://gerrit2.aldebaran.lan/561
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>