jni.jnitools: Corrects conversion between double and float. #46034
When we tried to set a double Property from java, it was casted in
float inside of the AnyReference pointing to its value.
Since float has a littlest capacity than double, this implied that
some value would not be able to be converted (for example
Double.MAX_VALUE).

This was due to an error in the signature creation from the property
type, that created a float signature for both double and float types.

Change-Id: I73ec013f3201673b7445d59df9130b71bd111c11
Reviewed-on: http://gerrit2.aldebaran.lan/629
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>